@@ -10,8 +10,10 @@
     "build": "npm run build:lib && npm run build:umd && npm run build:umd:min",
     "clean": "rimraf lib dist",
     "prepublish": "npm run clean && npm run build",
-    "lint": "eslint src",
-    "test": "NODE_ENV=eee mocha --compilers js:babel/register --recursive"
+    "lint": "eslint src develop",
+    "test": "NODE_ENV=eee mocha --compilers js:babel-register --recursive --require babel-polyfill",
+    "test:watch": "NODE_ENV=eee mocha --compilers js:babel-register --recursive  --require babel-polyfill --watch",
+    "start": "kotatsu serve --port 4000 --config ./develop/config/index.babel.js --presets es2015,stage-0,react,react-hmre ./develop/Main.js"
   },
   "repository": {
     "type": "git",
@@ -41,27 +43,46 @@
   "dependencies": {
     "eventemitter3": "^1.1.0",
     "fbjs": "^0.8.3",
-    "lodash": "^4.6.1",
     "point-geometry": "0.0.0",
     "scriptjs": "^2.5.7"
   },
   "devDependencies": {
-    "babel": "^5.8.23",
-    "babel-core": "^5.8.25",
+    "autoprefixer": "^6.3.6",
+    "babel-cli": "^6.10.1",
+    "babel-core": "^6.9.1",
     "babel-eslint": "^6.0.4",
-    "babel-loader": "^5.3.2",
+    "babel-loader": "^6.2.4",
+    "babel-plugin-webpack-loaders": "^0.5.0",
+    "babel-polyfill": "^6.9.1",
+    "babel-preset-es2015": "^6.9.0",
+    "babel-preset-react": "^6.5.0",
+    "babel-preset-react-hmre": "^1.1.1",
+    "babel-preset-stage-0": "^6.5.0",
+    "babel-register": "^6.9.0",
+    "css-loader": "^0.23.1",
     "eslint": "^2.12.0",
     "eslint-config-airbnb": "^9.0.1",
     "eslint-plugin-import": "^1.8.1",
     "eslint-plugin-jsx-a11y": "^1.4.2",
     "eslint-plugin-react": "^5.1.1",
     "expect": "^1.11.1",
+    "file-loader": "^0.8.5",
     "jsdom": "^6.5.1",
+    "kotatsu": "^0.14.0",
+    "lodash": "^4.6.1",
     "mocha": "^2.3.3",
-    "react": "^0.14.0",
-    "react-addons-test-utils": "^0.14.0",
-    "react-dom": "^0.14.0",
+    "node-sass": "^3.7.0",
+    "normalize.css": "^4.1.1",
+    "postcss-loader": "^0.9.1",
+    "react": "^15.1.0",
+    "react-addons-test-utils": "^15.1.0",
+    "react-dom": "^15.1.0",
+    "react-motion": "^0.4.4",
+    "recompose": "^0.19.0",
     "rimraf": "^2.4.3",
+    "sass-loader": "^3.2.0",
+    "style-loader": "^0.13.1",
+    "url-loader": "^0.5.7",
     "webpack": "^1.12.2"
   }
 }
